Adds FOGI analysis to cloud noise models; checks results with std 2Q explicit model.
Implements CloudNoiseModel.setup_fogi and fixes a number of bugs & issues
related to implicit-model FOGI analysis found by comparing a 2Q explicit model
with essentially the same 2Q cloud noise model.  Fixes included:

- using the *dual* error generator where appropriate when constructing
  "gauge action" matrices, instead of just the conjugate of the primal
  elementary error generator.
- fixing the assignment of elementary errorgen labels to gauge directions.
  Previously we were assuming the gauge-space directions were elementary
  errorgen directions, but actually `gauge_space.vectors` holds each gauge
  direction in terms of elementary directions.  Leaving out this middle step
  would cause normalization errors and other issues.

This commit adds and then comments out & tags with REMOVE numerous debug
print statements in fogitools.py and the model class files.  These should
be removed later.  Also, this commit creates separate setup_fogi implementations
in each of the LocalNoiseModel and CloudNoiseModel files -- this was primarily
for debugging and independent tweaking.  In the near future, we should try
to consolidate the setup_fogi routines, maybe even with ExplicitOpModel.setup_fogi.

Finally, note that no attempt to add FOGI parameterizations to implicit models
has been made yet -- this is a task for furture work.
